Trying to comment on what you wrote
let me know if I've understood it wrong.
I think the one flaw in this is repeat count. This would work if there only one set of each type of brackets.

As I said a while ago (because I've been shit at getting back to you! So so sorry!), this exercise is tricky because there is only one simple solution and anything beyond it is really convoluted.

The secret sauce to this exercise is a data structure called a Stack. In ruby, it'll be a list like any other but where, like a stack of plate, you can only add to the top (push()) or remove from the top (pop). A bit of a "last in first out" kind of model. It has the advantage of keeping a good "chronological" record of things. Things are always in order of seniority in it.

In the case of our problem, we know that:
- an open bracket always has to be closed,
- an open bracket cannot be closed if a *more recent* bracket has not been closed.
Which makes stacks very useful. Here's the solution in words and then in Python because I can't do Ruby).

You've got an empty stack and a string of brackets. For each bracket, you check if it opens, if it does you push it onto the stack. If it closes, you "pop" the latest element of the stack, check if it closes that latest element, if yes, you move on, if no, your string is wrong.
At the end, if your stack is not empty, your string is wrong.

``` python
# This will allow us to enter our test values
brackets = input("Please enter a list of brackets. >")

# we start with an empty stack
stack = []

# This will help us see if brackets open or if they close the right bracket
# If it is a key of the dictionary, it opens and if it is a value, we check the key to it
bracket_dictionary = {
    "{": "}",
    "[": "]",
    "(": ")"
}

valid = True

for bracket in brackets:
    # This checks if this bracket is a key in the dictionary, and therefore an opening bracket
    if bracket in bracket_dictionary:
        # add it to the stack, the equivalent in ruby is "push()"
        stack.append(bracket)
        # This "continue" key makes us skip the rest of the block and move on straight to the next iteration
        continue
    # If we get here, the bracket is a closing one
    # we "pop" the latest bracket off the stack and check if our current bracket closes it
    opening = stack.pop()
    if bracket_dictionary[opening] != bracket:
        valid = False
        # If this is wrong, the whole string is wrong so let's end it now
        # "break" ends the whole loop and sends us straight to line 34
        break

# If the length of the stack is not 0 at this point, there's a dangling bracket somewhere
if len(stack) != 0:
    valid = False

# Final message
if valid == True:
    print("Valid!")
else:
    print ("Invalid!")
```